requirements.txt
aiohttp==3.8.5
aiosignal==1.2.0
amadeus==9.0.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2023.7.22
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==41.0.3
defusedxml==0.7.1
Django==3.2.20
django-allauth==0.45.0
django-ckeditor==6.6.1
django-cors-headers==3.13.0
django-js-asset==2.1.0
django-sendgrid-v5==1.1.1
django-simple-coupons==0.5
frozenlist==1.2.0
geoip2==4.5.0
gunicorn==20.1.0
idna==3.3
ipinfo==3.0.0
ipinfo-django==1.1.0
maxminddb==2.2.0
multidict==5.2.0
mysqlclient==2.1.0
oauthlib==3.2.2
pycountry==20.7.3
pycparser==2.20
PyJWT==2.4.0
python-dotenv==0.14.0
python-http-client==3.3.3
python3-openid==3.2.0
pytz==2021.3
requests==2.31.0
requests-oauthlib==1.3.0
sendgrid==6.9.1
six==1.16.0
sqlparse==0.4.4
sslcommerz-lib==1.0
starkbank-ecdsa==2.0.2
stripe==2.61.0
twilio==7.12.0
typing_extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2